% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_positions.R
\name{grid_positions}
\alias{grid_positions}
\title{Generate a 2d grid in [0,1]^2 of given size.}
\usage{
grid_positions(n, m)
}
\arguments{
\item{n}{Integer giving one dimension of the grid.}

\item{m}{integer giving the other dimension of the grid.}
}
\value{
A (nm)x2 grid containing the positions of the desired grid.
}
\description{
Generates a matrix containing the positions of the points of an equidistant 2d grid in [0,1]^2.
}
