% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sinkhorn_divergence.R
\name{sinkhorn_divergence}
\alias{sinkhorn_divergence}
\title{Sinkhorn Divergence}
\usage{
sinkhorn_divergence(
  supplyList,
  demandList,
  supplyDivList,
  demandDivList,
  epsi = 1e-08,
  p = 2,
  exp = 1,
  ssCostMatrix = NULL,
  sdCostMatrix = NULL,
  ddCostMatrix = NULL,
  maxIteration = 5000,
  tol = 0.001
)
}
\arguments{
\item{supplyList}{A list containing the supply measure and the support of the supply distribution (see \link[WSGeometry]{regularized_transport}).}

\item{demandList}{A list containing the demand measure and the support of the demand distribution (see \link[WSGeometry]{regularized_transport}).}

\item{supplyDivList}{A list describing the supply divergence (see \link[WSGeometry]{regularized_transport}).}

\item{demandDivList}{A list describing the demand divergence (see \link[WSGeometry]{regularized_transport}).}

\item{epsi}{(optional) Final regularization strength used for approximation (default: \code{1e-8}).}

\item{p}{(optional) Norm to use for cost matrix calculation (default: 2).}

\item{exp}{(optional) Power to apply to calculated cost matrix (default: 1).}

\item{ssCostMatrix}{(optional) Cost matrix for transport between supply and itself, if \code{NULL} it is calculated from \code{supplyList[[2]]} (default: \code{NULL}).}

\item{sdCostMatrix}{(optional) Cost matrix for transport between supply and demand, if \code{NULL} it is calculated from \code{supplyList[[2]]} and \code{demandList[[2]]} (default: \code{NULL}).}

\item{ddCostMatrix}{(optional) Cost matrix for transport between demand and itself, if \code{NULL} it is calculated from \code{demandList[[2]]} (default: \code{NULL}).}

\item{maxIteration}{(optional) Maximal number of iterations to perform (default: \code{5000}).}

\item{tol}{(optional) convergence tolerance for Sinkhorn algorithm (default: \code{1e-3}).}
}
\value{
A single number: the calculated Sinkhorn divergence
}
\description{
Calculate the entropic Sinkhorn divergence for unbalanced transport
}
\details{
This function calculates the entropic Sinkhorn divergence for unbalanced transport defined as 
\eqn{OT_{\varepsilon}(\alpha, \beta) - \frac{1}{2} OT_{\varepsilon}(\alpha, \alpha) - \frac{1}{2} OT_{\varepsilon}(\beta, \beta) + \frac{\varepsilon}{2}(m(\alpha) - m(\beta))^2}{OT(a, b) - 0.5 OT(a, a) - 0.5 OT(b, b) - 0.5 (m(a) - m(b))^2}
where \eqn{m(\alpha)}{m(a)} is the total mass of \eqn{\alpha}{a}.
}
\examples{
supplyPoints <- seq(0, 1, length.out=100)
demandPoints <- seq(0, 1, length.out=100)

supply <- rep(0, 100)
supply[2:20] = 2
supply[91:100] = c(seq(0, 2, length.out=5), seq(2, 0, length.out=5))

demand <- rep(0, 100)
demand[21:40] = seq(0, 2, length.out=20)
demand[51:90] = 1.3 * sqrt(1 - seq(-1, 1, length.out=40)^2)

supplyList <- list(supply, supplyPoints)
demandList <- list(demand, demandPoints)

supplyDiv <- list("TV", 1, 1)
demandDiv <- list("TV", 1, 1)

epsi <- 1e-3

sinkhorn_divergence(supplyList, demandList, supplyDiv, demandDiv, 
                    epsi, maxIteration=10000, tol=1e-4)

}
\references{
Séjourné T, Feydy J, Vialard F, Trouvé A, Peyré G (2021). “Sinkhorn Divergences for Unbalanced Optimal Transport.” arXiv: 1910.12958. 1910.12958v2.
}
