% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfr.R
\name{wfr}
\alias{wfr}
\title{Wasserstein-Fisher-Rao Distance}
\usage{
wfr(
  supplyList,
  demandList,
  delta,
  epsi = 1e-08,
  maxIteration = 5000,
  tol = 0.001
)
}
\arguments{
\item{supplyList}{A list containing the supply measure and the support of the supply distribution (see example).}

\item{demandList}{A list containing the demand measure and the support of the demand distribution (see example).}

\item{delta}{Transport range parameter.}

\item{epsi}{(optional) Final regularization strength used for approximation (default: \code{1e-8}).}

\item{maxIteration}{(optional) Maximal number of iterations to perform (default: \code{5000}).}

\item{tol}{(optional) convergence tolerance for Sinkhorn algorithm (default: \code{1e-3}).}
}
\value{
A single number: the entropic WFR distance.
}
\description{
Approximate the Wasserstein-Fisher-Rao distance between two measures using entropic transport.
}
\details{
This function approximates the Wasserstein-Fisher-Rao distance between two measures using 
the Sinkhorn algorithm based on Corollary 5.9 in [1]. The theoretical exact result corresponds to the limit of \code{epsi} going to 0.
}
\examples{
supplyPoints <- seq(0, 1, length.out=100)
demandPoints <- seq(0, 1, length.out=100)

supply <- rep(0, 100)
supply[2:20] = 2
supply[91:100] = c(seq(0, 2, length.out=5), seq(2, 0, length.out=5))

demand <- rep(0, 100)
demand[21:40] = seq(0, 2, length.out=20)
demand[51:90] = 1.3 * sqrt(1 - seq(-1, 1, length.out=40)^2)

supplyList <- list(supply, supplyPoints)
demandList <- list(demand, demandPoints)

epsi <- 1e-3
delta <- 1

wfr(supplyList, demandList, delta, epsi)

}
\references{
[1] Lenaic Chizat, Gabriel Peyré, Bernhard Schmitzer, François-Xavier Vialard (2019). “Unbalanced Optimal Transport: Dynamic and Kantorovich Formulation” arXiv: 1508.05216v3.
}
