% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpp_rework.R
\name{wpp}
\alias{wpp}
\title{Constructor function for \link[transport]{wpp-object}s.}
\usage{
wpp(coordinates, mass)
}
\arguments{
\item{coordinates}{A Mxd matrix where each row represents one location of a support point of the point pattern.}

\item{mass}{A vector of length M of non-negative values specifying the mass at the given locations.}
}
\value{
A \link[transport]{wpp-object} corresponding to the given mass and coordinates.
}
\description{
Generates a \link[transport]{wpp-object} from given coordinates and weights.
}
