% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wppplot.R
\name{wppplot}
\alias{wppplot}
\title{Plot wpp-objects and transport plans between them.}
\usage{
wppplot(
  A,
  B = NULL,
  plan = NULL,
  p = NULL,
  C = NULL,
  pch = c(16, 16),
  bg = c("blue", "red"),
  col = c("blue", "red"),
  linecol = "black",
  cex = 1,
  bwd = 2,
  lwd = 2,
  xlim = c(0, 1),
  ylim = c(0, 1),
  clean = FALSE,
  tstart = "center",
  arrowScale = 0.03
)
}
\arguments{
\item{A}{A \link[transport]{wpp-object}.}

\item{B}{A \link[transport]{wpp-object}.}

\item{plan}{A matrix representing the OT plan between A and B.}

\item{p}{An integer specifying the power of the Euclidean distance used for computing the OT plan 
if none is provided.}

\item{C}{The penalty C of the Kantorovich-Rubinstein distance used for computing the UOT plan if no
plan is provided.}

\item{pch}{A graphical parameter specifying the shape of the locations in the plots. The first entry determined this
for A and the second for B.}

\item{bg}{A graphical parameter determining the colour of the locations for certain choices of pch. The first entry determined this
for A and the second for B.}

\item{col}{A graphical parameter specifying the colour of the locations or the colour of the border of 
the locations depending on pch. The first entry determined this
for A and the second for B.}

\item{linecol}{A graphical parameter specifying the colour of the transport plan.}

\item{cex}{A graphical parameter specifying the size of the plotted points.}

\item{bwd}{A graphical parameter specifying the size of the border of the plotted points when applicable.}

\item{lwd}{A graphical parameter specifying the width of the OT plans.}

\item{xlim}{A vector with two entries specifying the range of the plot on the x-axis.}

\item{ylim}{A vector with two entries specifying the range of the plot on the y-axis.}

\item{clean}{A graphical parameter specifying whether the axis are plotted.}

\item{tstart}{A graphical parameter specifying the visuals of the transport plans. The available options are "center",
"border" and "arrow". For "center" the lines start and end in the center of the points. For "border" they start and end 
at the border. For "arrow" the transport plan is represented by arrows and the start and end point are controlled by the 
parameter arrowScale.}

\item{arrowScale}{A real number specifying the starting and end positions of the transport plan when using the "arrow" 
parameter for tstart.}
}
\value{
This function has no return value and is only called for generating plots.
}
\description{
This function plots one or two wpp-objects and transport plans between them.
}
\examples{
M<-8
set.seed(76597)
pos1<-cbind(runif(M,0.1,0.8),runif(M,0.1,0.9))
pos2<-cbind(runif(M,0.1,0.9),runif(M,0.2,0.9))
W1<-rep(1,M)
W2<-rep(1,M)
wpp1<-wpp(pos1,W1)
wpp2<-wpp(pos2,W2)
wppplot(wpp1,wpp2,plan=transport::transport(wpp1,wpp2,p=2),cex=3,bwd=2,lwd=2,clean = TRUE
,tstart = "border",pch=c(21,21),col=c("black","black"))
wppplot(wpp1,wpp2,plan=transport::transport(wpp1,wpp2,p=2),cex=3,bwd=2,lwd=2,clean = TRUE
,tstart = "border",pch=c(22,25),col=c("black","black"),bg=c("green","orange"))
wppplot(wpp1,wpp2,plan=transport::transport(wpp1,wpp2,p=2),cex=3,bwd=2,lwd=2,clean = TRUE
,tstart = "center",pch=c(22,25),col=c("blue","red"),bg=c("green","orange"))
wppplot(wpp1,wpp2,plan=transport::transport(wpp1,wpp2,p=2),cex=3,bwd=2,lwd=2,clean = TRUE
,tstart = "arrow",pch=c(20,20),col=c("purple","brown"),bg=c("brown","purple"),arrowScale = 0.02)
wppplot(wpp1,wpp2,p=2,C=0.1,cex=3,bwd=2,lwd=2,clean = TRUE
,tstart = "border",pch=c(21,21),col=c("black","black"))
}
